+++ core/models/classifiers.go
@@ -50,9 +50,9 @@ func init() {
 	goflow.RegisterClassificationServiceFactory(classificationServiceFactory)
 }
 
-func classificationServiceFactory(c *runtime.Config) engine.ClassificationServiceFactory {
+func classificationServiceFactory(rt *runtime.Runtime) engine.ClassificationServiceFactory {
 	return func(classifier *flows.Classifier) (flows.ClassificationService, error) {
-		return classifier.Asset().(*Classifier).AsService(c, classifier)
+		return classifier.Asset().(*Classifier).AsService(rt.Config, classifier)
 	}
 }